ldrELFRelocatable.cpp.h: pointer calc warning fix for 32-bit.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@46180 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/ldr/ldrELFRelocatable.cpp.h
@@ -770,10 +770,11 @@ static DECLCALLBACK(int) RTLDRELF_NAME(EnumDbgInfo)(PRTLDRMODINTERNAL pMod, cons
 
             RT_ZERO(DbgInfo.u);
             DbgInfo.enmType         = RTLDRDBGINFOTYPE_DWARF_DWO;
-            DbgInfo.pszExtFile      = (const char *)((uintptr_t)pModElf->pvBits + paShdrs[iShdr].sh_offset);
+            DbgInfo.pszExtFile      = (const char *)((uintptr_t)pModElf->pvBits + (uintptr_t)paShdrs[iShdr].sh_offset);
             if (!RTStrEnd(DbgInfo.pszExtFile, paShdrs[iShdr].sh_size))
                 return VERR_BAD_EXE_FORMAT;
-            DbgInfo.u.Dwo.uCrc32    = *(uint32_t *)((uintptr_t)DbgInfo.pszExtFile + paShdrs[iShdr].sh_size - sizeof(uint32_t));
+            DbgInfo.u.Dwo.uCrc32    = *(uint32_t *)((uintptr_t)DbgInfo.pszExtFile + (uintptr_t)paShdrs[iShdr].sh_size
+                                                    - sizeof(uint32_t));
             DbgInfo.offFile         = -1;
             DbgInfo.cb              = 0;
         }
